Refactor KMC module structure and update imports
- Moved KMC class from kmcpy/kmc.py to kmcpy/simulation/kmc.py for better organization.
- Updated import paths throughout the codebase to reflect the new location of the KMC class.
- Changed execution count in nasicon.ipynb to null for consistency.
- Adjusted test files to import KMC from the new module path.
- Removed the old kmc.py file to clean up the codebase.
